+++ CHANGELOG.rst
@@ -6,7 +6,9 @@ Changelog
 
 Features:
 
+* Add ``FalconParser`` (:issue:`63`).
 * Add ``fields.DelimitedList`` (:issue:`66`). Thanks :user:`jmcarp`.
+* ``TornadoParser`` will parse json with ``simplejson`` if it is installed.
 
 0.17.0 (2015-09-29)
 *******************